Refactor order management functions into a new orderMgmt module and update imports in vendor orders page
